refactor: вынесяна в отдельный класс валидация
