+++ src/main/java/ru/yandex/practicum/filmorate/service/DirectorService.java
@@ -2,8 +2,6 @@ package ru.yandex.practicum.filmorate.service;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
-import ru.yandex.practicum.filmorate.exceptions.IllegalIdException;
-import ru.yandex.practicum.filmorate.exceptions.InvalidDirectorException;
 import ru.yandex.practicum.filmorate.exceptions.StorageException;
 import ru.yandex.practicum.filmorate.model.Director;
 import ru.yandex.practicum.filmorate.storage.dao.DirectorDao;
@@ -15,10 +13,12 @@ import java.util.Optional;
 public class DirectorService {
 
     private final DirectorDao directorDao;
+    private final ValidationService validationService;
 
     @Autowired
-    public DirectorService(DirectorDao directorDao) {
+    public DirectorService(DirectorDao directorDao, ValidationService validationService) {
         this.directorDao = directorDao;
+        this.validationService = validationService;
     }
 
     public List<Director> get() {
@@ -26,38 +26,27 @@ public class DirectorService {
     }
 
     public Director getDirectorById(long directorId) {
-        checkId(directorId);
+        validationService.checkId(directorId);
         Optional<Director> director = directorDao.findDirectorById(directorId);
         if (director.isEmpty()) throw new StorageException(String.format("Режиссёр с id: %d не найден", directorId));
         return director.get();
     }
 
     public Director add(Director director) {
-        validateDirector(director);
+        validationService.validateDirector(director);
         return directorDao.save(director);
     }
 
     public Director update(Director director) {
         checkIfDirectorExists(director.getId());
-        validateDirector(director);
+        validationService.validateDirector(director);
         return directorDao.update(director);
     }
 
     public void delete(long directorId) {
-        checkId(directorId);
+        validationService.checkId(directorId);
         checkIfDirectorExists(directorId);
         directorDao.delete(directorId);
-    }
-
-    private void validateDirector(Director director) {
-        if (director.getId() <= 0
-                || director.getName() == null
-                || director.getName().isBlank()
-        ) throw new InvalidDirectorException("Недопустимы значения поле director");
-    }
-
-    private void checkId(long id) {
-        if (id < 0) throw new IllegalIdException("У режиссёра не может быть отрицательный id");
     }
 
     public void checkIfDirectorExists(long directorId) {
