+++ src/main/java/ru/yandex/practicum/filmorate/service/FilmService.java
@@ -1,11 +1,8 @@
 package ru.yandex.practicum.filmorate.service;
 
-import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.stereotype.Service;
-import ru.yandex.practicum.filmorate.exceptions.IllegalIdException;
-import ru.yandex.practicum.filmorate.exceptions.InvalidFilmException;
 import ru.yandex.practicum.filmorate.model.Feed;
 import ru.yandex.practicum.filmorate.model.Film;
 import ru.yandex.practicum.filmorate.sorts.SortingType;
@@ -20,33 +17,34 @@ import java.util.List;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
-@Slf4j
 @Service
 public class FilmService {
-    public static final long MAX_FILM_DESCRIPTION_LENGTH = 200L;
-    public static final LocalDate FIRST_FILM_BIRTHDAY = LocalDate.of(1895, Month.DECEMBER, 28);
 
     private final FilmDao filmDao;
     private final LikeDao likeDao;
-    private final DirectorService directorService;
     private final DbFeedDaoImpl dbFeedDaoImpl;
+    private final DirectorService directorService;
+    private final ValidationService validationService;
 
     @Autowired
     public FilmService(@Qualifier("dbFilmDaoImpl") FilmDao filmDao,
-                       LikeDao likeDao, DirectorService directorService, DbFeedDaoImpl dbFeedDaoImpl) {
+                       LikeDao likeDao, DirectorService directorService,
+                       DbFeedDaoImpl dbFeedDaoImpl,
+                       ValidationService validationService) {
         this.filmDao = filmDao;
         this.likeDao = likeDao;
         this.directorService = directorService;
         this.dbFeedDaoImpl = dbFeedDaoImpl;
+        this.validationService = validationService;
     }
 
     public Film add(Film film) {
-        validate(film);
+        validationService.validate(film);
         return filmDao.save(film);
     }
 
     public Film update(Film film) {
-        validate(film);
+        validationService.validate(film);
         return filmDao.update(film);
     }
 
@@ -55,7 +53,7 @@ public class FilmService {
     }
 
     public Film getFilmById(long id) {
-        checkFilmId(id);
+        validationService.checkFilmId(id);
         return filmDao.findFilmById(id).orElse(null);
     }
 
@@ -66,16 +64,16 @@ public class FilmService {
     }
 
     public void addLike(long filmId, long userId) {
-        checkFilmId(filmId);
-        checkUserId(userId);
+        validationService.checkFilmId(filmId);
+        validationService.checkUserId(userId);
         dbFeedDaoImpl.saveFeed(new Feed(1, Instant.now().toEpochMilli(),
                 userId, "LIKE", "ADD", filmId));
         likeDao.addLike(filmId, userId);
     }
 
     public void deleteLike(long filmId, long userId) {
-        checkFilmId(filmId);
-        checkUserId(userId);
+        validationService.checkFilmId(filmId);
+        validationService.checkUserId(userId);
         dbFeedDaoImpl.saveFeed(new Feed(1, Instant.now().toEpochMilli(),
                 userId, "LIKE", "REMOVE", filmId));
         likeDao.deleteLike(filmId, userId);
@@ -104,66 +102,13 @@ public class FilmService {
     }
 
     public void deleteFilmById(Long filmId) {
-        checkFilmId(filmId);
+        validationService.checkFilmId(filmId);
         filmDao.deleteFilmById(filmId);
     }
 
     public List<Film> getDirectorFilms(long directorId, SortingType sortBy) {
-        checkDirectorId(directorId);
+        validationService.checkDirectorId(directorId);
         directorService.checkIfDirectorExists(directorId);
         return filmDao.getDirectorFilms(directorId, sortBy);
-    }
-
-    private void validate(Film film) {
-        validateNotNull(film);
-        checkFilmId(film.getId());
-        if (film.getName() == null) {
-            String message = "Объект Film некорректно инициализирован, есть null поля! id: " + film.getId();
-            log.warn(message);
-            throw new InvalidFilmException(message);
-        }
-        if (film.getName().isBlank()) {
-            String message = "Пустое имя фильма при инициализации id: " + film.getId();
-            log.warn(message);
-            throw new InvalidFilmException(message);
-        }
-        if (film.getDescription() != null && film.getDescription().length() > MAX_FILM_DESCRIPTION_LENGTH) {
-            String message = "Описание длиннее " + MAX_FILM_DESCRIPTION_LENGTH + " id: " + film.getId();
-            log.warn(message);
-            throw new InvalidFilmException(message);
-        }
-        if (film.getReleaseDate() != null && film.getReleaseDate().isBefore(FIRST_FILM_BIRTHDAY)) {
-            String message = "Дата релиза раньше ДР кино:" + FIRST_FILM_BIRTHDAY + " id: " + film.getId();
-            log.warn(message);
-            throw new InvalidFilmException(message);
-        }
-        if (film.getDuration() < 0) {
-            String message = "Отрицательная продолжительность фильма  id: " + film.getId();
-            log.warn(message);
-            throw new InvalidFilmException(message);
-        }
-        if (film.getMpa() == null) {
-            throw new InvalidFilmException(String.format("у фильма id: %d не установлен mpa", film.getId()));
-        }
-    }
-
-    private void validateNotNull(Film film) {
-        if (film == null) {
-            String message = "Передан null film";
-            log.warn(message);
-            throw new IllegalStateException(message);
-        }
-    }
-
-    public void checkFilmId(long id) {
-        if (id < 0) throw new IllegalIdException("film id:" + id + " отрицательный");
-    }
-
-    public void checkUserId(long id) {
-        if (id < 0) throw new IllegalIdException("user id: " + id + " отрицательный");
-    }
-
-    public void checkDirectorId(long id) {
-        if (id < 0) throw new IllegalIdException("director id: " + id + " отрицательный");
     }
 }
