+++ src/main/java/ru/yandex/practicum/filmorate/service/GenreService.java
@@ -2,7 +2,6 @@ package ru.yandex.practicum.filmorate.service;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
-import ru.yandex.practicum.filmorate.exceptions.IllegalIdException;
 import ru.yandex.practicum.filmorate.model.Genre;
 import ru.yandex.practicum.filmorate.storage.dao.GenreDao;
 
@@ -11,10 +10,12 @@ import java.util.List;
 @Service
 public class GenreService {
     private final GenreDao genreDao;
+    private final ValidationService validationService;
 
     @Autowired
-    public GenreService(GenreDao genreDao) {
+    public GenreService(GenreDao genreDao, ValidationService validationService) {
         this.genreDao = genreDao;
+        this.validationService = validationService;
     }
 
     public List<Genre> findAllGenres() {
@@ -22,11 +23,7 @@ public class GenreService {
     }
 
     public Genre findGenreById(long id) {
-        checkId(id);
+        validationService.checkId(id);
         return genreDao.findGenreById(id);
-    }
-
-    private void checkId(long id) {
-        if (id < 0) throw new IllegalIdException("У жанра не может быть отрицательный id");
     }
 }
