+++ src/main/java/ru/yandex/practicum/filmorate/service/MpaService.java
@@ -2,7 +2,6 @@ package ru.yandex.practicum.filmorate.service;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
-import ru.yandex.practicum.filmorate.exceptions.IllegalIdException;
 import ru.yandex.practicum.filmorate.model.Mpa;
 import ru.yandex.practicum.filmorate.storage.dao.MpaRatingDao;
 
@@ -12,10 +11,12 @@ import java.util.List;
 public class MpaService {
 
     private final MpaRatingDao mpaRatingDao;
+    private final ValidationService validationService;
 
     @Autowired
-    public MpaService(MpaRatingDao mpaRatingDao) {
+    public MpaService(MpaRatingDao mpaRatingDao, ValidationService validationService) {
         this.mpaRatingDao = mpaRatingDao;
+        this.validationService = validationService;
     }
 
     public List<Mpa> findAllMpaRatings() {
@@ -23,11 +24,7 @@ public class MpaService {
     }
 
     public Mpa findMpaRatingById(long id) {
-        checkId(id);
+        validationService.checkId(id);
         return mpaRatingDao.findMpaRatingById(id);
-    }
-
-    private void checkId(long id) {
-        if (id < 0) throw new IllegalIdException("mpa id не может быть отрицательным");
     }
 }
