+++ src/main/java/ru/yandex/practicum/filmorate/service/UserService.java
@@ -5,7 +5,6 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.stereotype.Service;
 import ru.yandex.practicum.filmorate.exceptions.IllegalIdException;
-import ru.yandex.practicum.filmorate.exceptions.InvalidUserException;
 import ru.yandex.practicum.filmorate.model.Feed;
 import ru.yandex.practicum.filmorate.model.Film;
 import ru.yandex.practicum.filmorate.model.User;
@@ -15,7 +14,6 @@ import ru.yandex.practicum.filmorate.storage.dao.UserDao;
 import ru.yandex.practicum.filmorate.storage.impl.DbFeedDaoImpl;
 
 import java.time.Instant;
-import java.time.LocalDate;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
@@ -25,28 +23,30 @@ import java.util.stream.Collectors;
 @Service
 public class UserService {
     private final UserDao userDao;
-    private final FriendshipDao friendshipDao;
     private final FilmDao filmDao;
     private final DbFeedDaoImpl feedDaoImpl;
-
+    private final FriendshipDao friendshipDao;
+    private final ValidationService validationService;
 
     @Autowired
     public UserService(@Qualifier("dbUserDaoImpl") UserDao userDao,
                        FriendshipDao friendshipDao,
-                       FilmDao filmDao, DbFeedDaoImpl feedDaoImpl) {
+                       FilmDao filmDao, DbFeedDaoImpl feedDaoImpl,
+                       ValidationService validationService) {
         this.userDao = userDao;
         this.friendshipDao = friendshipDao;
         this.filmDao = filmDao;
         this.feedDaoImpl = feedDaoImpl;
+        this.validationService = validationService;
     }
 
     public User add(User user) {
-        validate(user);
+        validationService.validate(user);
         return userDao.save(user);
     }
 
     public User update(User user) {
-        validate(user);
+        validationService.validate(user);
         return userDao.update(user);
     }
 
@@ -55,27 +55,27 @@ public class UserService {
     }
 
     public User getUserById(long id) {
-        checkNegativeIds(id);
+        validationService.checkNegativeIds(id);
         return userDao.findUserById(id)
                 .orElseThrow(() -> new IllegalIdException(String.format("Пользователь %d не найден", id)));
     }
 
     public void addFriend(long userId, long friendId) {
-        checkNegativeIds(userId, friendId);
+        validationService.checkNegativeIds(userId, friendId);
         feedDaoImpl.saveFeed(new Feed(1, Instant.now().toEpochMilli(),
                 userId,"FRIEND","ADD", friendId));
         friendshipDao.addFriend(userId, friendId);
     }
 
     public void deleteFriend(long userId, long friendId) {
-        checkNegativeIds(userId, friendId);
+        validationService.checkNegativeIds(userId, friendId);
         feedDaoImpl.saveFeed(new Feed(1, Instant.now().toEpochMilli(),
                 userId,"FRIEND","REMOVE", friendId));
         friendshipDao.deleteFriend(userId, friendId);
     }
 
     public List<User> getUserFriends(long id) {
-        checkNegativeIds(id);
+        validationService.checkNegativeIds(id);
         List<User> result = new ArrayList<>();
         Optional<User> optionalUser = userDao.findUserById(id);
         if (optionalUser.isPresent()) {
@@ -88,7 +88,7 @@ public class UserService {
     }
 
     public List<User> getCommonFriends(long user1Id, long user2Id) {
-        checkNegativeIds(user1Id, user2Id);
+        validationService.checkNegativeIds(user1Id, user2Id);
         List<User> commonFriends = new ArrayList<>();
         List<Long> user1FriendsId = friendshipDao.getFriends(user1Id);
         List<Long> user2FriendsId = friendshipDao.getFriends(user2Id);
@@ -101,12 +101,12 @@ public class UserService {
     }
 
     public void deleteUserById(Long userId) {
-        checkNegativeIds(userId);
+        validationService.checkNegativeIds(userId);
         userDao.deleteUserById(userId);
     }
 
     public List<Film> recommendationsFilms(Long id) {
-        checkNegativeIds(id);
+        validationService.checkNegativeIds(id);
         List<Film> userFilms = new ArrayList<>(filmDao.findAllFavoriteMovies(id));
         List<Film> recommendationsFilms = new ArrayList<>(filmDao.recommendationsFilm(id));
         return recommendationsFilms.stream()
@@ -114,55 +114,7 @@ public class UserService {
                 .collect(Collectors.toList());
     }
 
-    public void checkNegativeIds(long... ids) {
-        for (long id : ids) {
-            if (id <= 0 ) throw new IllegalIdException("user id:" + id + " отрицательный");
-        }
-    }
-
     public List<Feed> getUserFeedList(Long userId){
         return feedDaoImpl.getUserFeedList(userId);
-    }
-
-    protected void validate(User user) {
-        validateNotNull(user);
-        if(user.getId() < 0) {
-            throw new IllegalIdException("id пользователя не может быть отрицательным");
-        }
-        if (user.getEmail() == null
-                || user.getLogin() == null
-                || user.getBirthday() == null
-        ) {
-            String message = "Некорректно инициализирован пользователь, есть null поля id: " + user.getId();
-            log.warn(message);
-            throw new NullPointerException(message);
-        }
-        if (user.getEmail().isBlank() || !user.getEmail().contains("@")) {
-            String message = "Некорректный адрес email id: " + user.getId();
-            log.warn(message);
-            throw new InvalidUserException(message);
-        }
-        if (user.getLogin().isBlank() || user.getLogin().contains(" ")) {
-            String message = "Логин пустой или содержит пробелы id: " + user.getId();
-            log.warn(message);
-            throw new InvalidUserException(message);
-        }
-        if (user.getName() == null || user.getName().isBlank()) {
-            user.setName(user.getLogin());
-            log.debug("Пользователю присвоено имя: " + user.getName());
-        }
-        if (user.getBirthday().isAfter(LocalDate.now())) {
-            String message = "День рождения указан в будущем id: " + user.getId();
-            log.warn(message);
-            throw new InvalidUserException(message);
-        }
-    }
-
-    private void validateNotNull(User user) {
-        if (user == null) {
-            String message = "Передан null user";
-            log.warn(message);
-            throw new IllegalStateException(message);
-        }
     }
 }
